Notification tile long press: show a persistent expanded vol panel
when long pressing on the tile, and remove the now useless arrow.
Removed the "return convertView" in the tile code so the volume
panel sliders in the tile detailed view are refreshed when the
user change the link/unlink notification/ringtone feature.

PS1: better code parsing for the last commit

Change-Id: Ic8cde6f4acff03b93bcad2c3cdc7f39b0f3b83b2
Signed-off-by: ROM-Jeremy <Disciple97414@Gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/volume/VolumePanel.java